Tests bootimage

+++ qscripts/build-riscv64.sh
@@ -9,8 +9,10 @@ cargo brv
 # assemble assembly
 $AS -c support/arch/riscv64/asm/entry.S -o build/entry.o
 # link objects
-$LD -T support/arch/riscv64/asm/linker.ld -nostdlib build/*.o build/rust/*.a -o build/kernel.elf
+$LD -T support/arch/riscv64/link/linker.ld -nostdlib build/*.o build/rust/*.a -o build/kernel.elf
 
 # clean up
+if [[ $# -eq 1 ]]; then
 rm -rf build/rust
 rm -rf build/entry.o
+fi
